+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features
 
 - popup: Improved Discogs Search button. Added more options for searching
-- popup: Added Bandcamp release links to the release card navigation
+- popup: Added Bandcamp release and search links to the release card navigation
 - popup: Output album styles on the album card and Bandcamp latest viewed widget
 
 ## 0.18.1 (2024-12-25)